Develop

+++ src/Components/Cards/ProductCard/ProductCard.jsx
@@ -5,7 +5,7 @@ import { useDispatch, useSelector } from 'react-redux'
 import { addToCart, removeFromCart, removeWish, addWish} from '../../../redux/actions.js' // CREAR UNA ACTION QUE DEPLOYE FAVORITO AL USUARIO
 import './ProductCard.css'
 import FavouriteButton from '../../FavouriteButton/FavouriteBurron.jsx'
-
+import Swal from 'sweetalert2'
 
 
 export default function ProductCard({ id, name, img, rating, platforms, price, inStock}) {
@@ -13,7 +13,13 @@ export default function ProductCard({ id, name, img, rating, platforms, price, i
   let foundCart=false;   //aca encontraria el juego si esta agregado al carrito
   const dispatch = useDispatch()
   
-  
+  const swalWithBootstrapButtons = Swal.mixin({
+    customClass: {
+      confirmButton: 'btn btn-success',
+      cancelButton: 'btn btn-danger'
+    },
+    buttonsStyling: false
+  })
   
   const handleClick = (e) => {
     e.preventDefault();
@@ -23,9 +29,42 @@ export default function ProductCard({ id, name, img, rating, platforms, price, i
         alert("Juego ya agregado al carrito anteriormente!")
        }else{
          dispatch(addToCart(id)) // dispacha al carrito de compras con el id del game en la db
+          Swal.fire({
+            position: 'center',
+            icon: 'success',
+            title: 'Succesfully added to your cart',
+            showConfirmButton: false,
+            timer: 1500
+          })
        }
     }else if(e.target.value==="remove"){
-      dispatch(removeFromCart(id))
+      swalWithBootstrapButtons.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'Yes, delete it!',
+        cancelButtonText: 'No, cancel!',
+        reverseButtons: true
+      }).then((result) => {
+        if (result.isConfirmed) {
+          swalWithBootstrapButtons.fire(
+            'Deleted!',
+            'Your product has been deleted from the cart.',
+            'success'
+          );
+          dispatch(removeFromCart(id))
+        } else if (
+          /* Read more about handling dismissals below */
+          result.dismiss === Swal.DismissReason.cancel
+        ) {
+          swalWithBootstrapButtons.fire(
+            'Cancelled',
+            'Your product is still in the cart ',
+            'error'
+          )
+        }
+      })
     }
 }
 
@@ -40,9 +79,8 @@ export default function ProductCard({ id, name, img, rating, platforms, price, i
           </Link>
           <div class="card-body" >
             <Link to={`/detail/${id}`} style={{ textDecoration: "none" }}>
-              <h5 class="card-title">{name}</h5>
+              <h6 class="card-title">{name}</h6>
             </Link>
-          
             {/* <div class="ratings">
               <svg aria-hidden="true" class="fa fa-star rating-color" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><title>First star</title><path d="M9.049 2.927c.3-.921 1.603-.921 1.902 0l1.07 3.292a1 1 0 00.95.69h3.462c.969 0 1.371 1.24.588 1.81l-2.8 2.034a1 1 0 00-.364 1.118l1.07 3.292c.3.921-.755 1.688-1.54 1.118l-2.8-2.034a1 1 0 00-1.175 0l-2.8 2.034c-.784.57-1.838-.197-1.539-1.118l1.07-3.292a1 1 0 00-.364-1.118L2.98 8.72c-.783-.57-.38-1.81.588-1.81h3.461a1 1 0 00.951-.69l1.07-3.292z"></path></svg>
               <svg aria-hidden="true" class="fa fa-star rating-color" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><title>Second star</title><path d="M9.049 2.927c.3-.921 1.603-.921 1.902 0l1.07 3.292a1 1 0 00.95.69h3.462c.969 0 1.371 1.24.588 1.81l-2.8 2.034a1 1 0 00-.364 1.118l1.07 3.292c.3.921-.755 1.688-1.54 1.118l-2.8-2.034a1 1 0 00-1.175 0l-2.8 2.034c-.784.57-1.838-.197-1.539-1.118l1.07-3.292a1 1 0 00-.364-1.118L2.98 8.72c-.783-.57-.38-1.81.588-1.81h3.461a1 1 0 00.951-.69l1.07-3.292z"></path></svg>
@@ -53,11 +91,20 @@ export default function ProductCard({ id, name, img, rating, platforms, price, i
             </div> */}
 
             {/* <input type="image" onClick={(e)=>handleClick(e)} value="favourite"  src={heart} class="m-2" style={{width:"2vw", filter:`${brigthness}`}}  alt="heart"/> */}
-            <FavouriteButton id={id}/>
-
-            {price?<span class="card-text">{price} U$d</span>:<span>No stock</span>}
-            <button disabled={price?false:true} onClick={(e) => handleClick(e)} value="cart" class="btn btn-primary">Carrito</button>
-            {foundCart&&<button onClick={(e) => handleClick(e)} type="button" class="btn-close" value="remove" aria-label="Close"></button>}
+            <div class="d-flex flex-row align-items-center justify-content-center">
+                <FavouriteButton id={id}/>
+                <div>
+                  {price?
+                  <span class="card-text bg-secondary m-2 p-2 text-light">
+                  ${price}
+                  </span>
+                  :<span>No stock</span>}
+                </div>
+                <div>
+                  <button disabled={price?false:true} onClick={(e) => handleClick(e)} value="cart" class="btn btn-primary">Cart</button>
+                </div>
+              {foundCart&&<button onClick={(e) => handleClick(e)} type="button" class="btn-close" value="remove" aria-label="Close"></button>}
+            </div>
           </div>
         </div>
